Implement case, for, and while structures in analog functions
Wnen I merged the feature-bug1168son branch on Friday, it was missing
support for case, for, and while structures in analog functions.  I am
now completing that implementation.

This also required fixing the code so it properly emits block-local
variables and doesn't try to differentiate integer variables.  We'll
only generate derivatives (and declarations for variables to hold
them) when we're working with real variables.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168