Remove another unnecessary indent command
BSIMCMG 110 requires post-ADMS code patching, but the patch is such
that it is NOT necessary to run emacs over it to reindent it first ---
it sufficient to use the "-l" (ignore whitespace) option to patch and
only indent the .h file.  BSIMCMG's .C file is so huge that indenting
it with emacs can take more than an hour, which is just pointless.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168