Add special lapack/basker embeddsampling solvers
I originally implemented embedded sampling to use ordering that was
easy to set up, but possibly not the most efficient in terms of the
linear solve.  That ordering was one in which the parameters were the
outer loop and the circuit unknowns were the inner loop.  This resulted
in a block linear system where the only nonzero blocks were on the block
diagonal, and those blocks each had the structure of the original
circuit matrix.

Another way to order things is to have the outer loop be circuit
unknowns and the inner loop be the parameters.  If done in that manner,
then the block structure of the block Jacobian matches that of the
original circuit matrix, and inside of each block is a diagonal matrix.
By ordering things in this manner, one can use solvers such as Basker,
which is designed to allow block matrix entries.

I have now implemented this second form of ordering for 3 solver
options: Lapack, regular Basker, and block Basker.  This implementation
is 99% inside of the N_LAS_ESDirectSolver.C file.  It does not attempt
to do the device evaluations differently, although I considered doing it
that way.  So, the original loaders and builders for embedded sampling
are still used, and they create and load a block matrix using the
original ordering.  Then, at matrix solver time, the contents of the
original block matrix are copied into new matrix structures for use by
either Lapack or Basker.  All of this copying and reordering is
performed in the ESDirectSolver class.

Writing it this way allowed me to copy a lot of code from the
HBDirectSolver class, which was written to do something very similar for
HB.

I also implemented a Lapack solve using the original ordering as a first
step for debugging.  So, in total, the specialized solver options are
now:
1) lapack using original ordering (params outer loop)
2) lapack using new ordering (params inner loop)
3) regular Basker using new ordering
4) block Basker using new ordering

These options are specified using:

* setting the new ordering:
.options EMBEDDEDSAMPLES PARAMSOUTERLOOP=false

* specifying the solver:
.options LINSOL-ES type=direct  direct_solver=basker

This somewhat follows the pattern of the specialized HB direct linear
solver.

As of this writing I am still debugging the 2 Basker options.  So,
it isn't quite done, but enough files got changed recently that
I was getting worried about losing my work.  Hence this commit.

Most of this work is in anticipation of the fully-intrusive PCE
analysis.  I think for that analysis I will probably use an approach
much like this one.  ie, keep the ES loads the same, but then (after
modification for PCE) reorder
and copy the results into a specialized linear solver strucrtures with
optimal ordering.  Doing it that was probably is the most expeditious
path forward for now.

This commit also includes a lot of small changes that I don't have time
to fully describe.

I went down several paths before settling on this one, and went far
enough to start writing code on at least one of the other paths.  I've
attempted to exclude all of that stuff from this commit (or remove it,
if it came in a previous commit) but it is possible I missed a thing or
two.