Make code actually use configure/cmake symbols
N_UTL_MathSpecialFunctions.C had some ifdefs such as "HAVE_ERFC" and
HAVE_ERF meant to make it use a system-provided erfc function if it
were available, and a substitute function if not.  HAVE_ERF and
HAVE_ERFC are created at configure time by a probe that checks if the
function exist, and these #defines are put in Xyce_config.h.

Unfortunately, when this file was created, it never got an include of
Xyce_config.h.  Therefore, "#ifdef HAVE_ERFC/use erfc/#else/use local
coded version" was always choosing the local coded version.

Originally, this file used "HAVE_ERF" as the guard for calling erfc,
and later this was corrected to "HAVE_ERFC" since the guard macro
should correspond to the probe that actually tested the function being
called.

Unfortunately, CMakeLists.txt (used only for Windows at the moment)
was *ALSO* wrong, because it set "HAVE_ERF" when EITHER erf or erfc
was found, and never set HAVE_ERFC even if erfc were found.  This
means that even if Xyce_config.h were included, we would still always
be choosing the local version of erfc.

This commit makes N_UTL_MathSpecialFunctions.C include Xyce_config.h
as it always should have done, and corrects CMakeLists.txt so that the
correct HAVE_ERFC macro is actually set when erfc is found.

I noticed this issue while considering how to support "hypot" as a
special function, given that 99% of all OSen and C++ compilers will
support it, but technically it's only required by C++11.  I was going
to ad a configure probe and make some conditional stuff in this file
just like we do for erfc, and then noticed that it wasn't correctly
set up to USE configure/cmake probe results anyway.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1227