Add "hypot" special function
"hypot" is a C99-standard math library function that computes
sqrt(x*x+y*y) with careful avoidance of under- and overflow.  Though
it's been in the C standard since C99, it only got into the C++
standard with C++11.  However, almost every single C++ compiler in the
world already supported it before C++11.

This commit adds a configure and CMake check for hypot, and a
functionn to N_UTL_MathSpecialFunctions.[Ch] called "hypot" that
either calls std::hypot (if it exists) or sqrt(x*x+y*y) if not.

This is in preparation to working on bug 1227.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1227