Revert "Fix typo in previous commit"
This reverts commit 972c0f04eb91ada3ffab8325a9650a79fc2ee368.

I have decided not to use the system-provided hypot function at all to
satisfy bug 1227, because that would break use of the function in
analog functions, where we are still making use of templated functions
that *may* wind up being instantiated with Sacado types (until we get
rid of Sacado even for sensitivity).

So I no longer need either the new special function in
N_UTL_MathSpecialFunctions OR the configure test for hypot.