Add support for Verilog-A "hypot" function
hypot(x,y) returns sqrt(x*x+y*y).  This commit adds support for this
function to the new Xyce/ADMS back-end (not backporting it to the
legacy back-end).

Tested both by using the function directly in an analog code block
and indirectly through an analog function.  Since no models use this,
I constructed a toy to use it and visually inspected the generated
code to assure correctness.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1227