Remove "std::" namespace for erf and erfc
Since N_UTL_MathSpecialFunctions was created, it was not properly set
up to include "Xyce_config.h", but was attempting to use HAVE_ERF and
HAVE_ERFC macros that would only have been found there to decide
whether to use a system-provided erf or erfc function.  Because of
this, it was NEVER using the function.

I "fixed" that earlier today, only to find every build on RHEL6
failing --- that's because while configure found erf and erfc, the
functions are NOT in the std:: namespace, and all the calls are
written as std::erf or std::erfc.

I'm removing the std:: namespace from the calls.  I have tested this
on RHEL6, RHEL7, and on BSD with both clang and gcc8, and this
compiles in all four builds.  I am hoping that it will not break on
Losedows, but odds are pretty good.