Fix incorrect ifdefs for losedows
The N_UTL_MathSpecialFunctions.C file had:
   #if (HAVE_ERFC)

and apparently when HAVE_ERFC is  defined by CMake it's not actually
given a value the way it is in autoconf.  Therefore the use of "#if"
instead of "#ifdef" results in a compile failure on Windows now that
HAVE_ERFC is being defined.

This commit changes #if to #ifdef.