Turn use of system-provided erfc back off
In an earlier commit, I "fixed" N_UTL_MathSpecialFunctions.C so that
it correctly used "erfc" and "erf" if the system libraries provide
it.  Unfortunately, this led to one LTRA test case hanging
mysteriously on two platforms (RHEL6 and Ubuntu 16.04) but not on any
others.

I have opened bug 1228 for this issue, but until "someone" has a
chance to hunt down the reason for the hang, I'm commenting out the
code that selects system erfc and erf and going back to the old
behavior of never using them that was in force since the file was
created in commmit  8e31961de with its error that basically made the
ifdef a no-op.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1228