Remove a file that never should have been committed
I created a file called "NOTES" in this directory that was never meant
to be part of the repo, and accidentally added/committed it.  This
causes problems for Mac users because there was already a file called
"Notes" in this directory, and Macs have case-insensitive file
systems.  grumble.

I am removing the file, but there exists the possibility that it will
not fix the problem completely for a Mac user with an existing
clone (because the git rm'd file might still get cornfused for the one
that isn't rm'd).  We might have to play around on the clone end to
clean up the mess.

I'll work with others to try to get this fixed on that end now that
I've fixed it on the repo end.