Update intrusive PCE to match new ES solvers
At this stage of development intrusive PCE is just a copy of embedded
sampling (ES) with different names, and with the regression PCE stuff
deleted.  Other than that, there is not difference.

So, when I updated ES to have specialized direct solvers, I needed to do
a similar update to PCE.  This commit accomplishes that.

Also does a tiny bit of cleanup in some of the ES files, mostly
correcting comments.