Initial Commit for YLIN Device
This is the initial commit for the YLIN device.  The main capability
is the ability to parse (Xyce-generated at least) Touchstone 2
files into private variables in a YLin::Model object.  The resultant
YLIN instances then act like a simplified R device with an R-value
equal to the first element in the impedance values listed on the
[Reference] (or Option) line in the Touchstone 2 file.  This will
allow further development work on the "device physics" of the YLIN
device.

SON Bug 1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215