add code to set up PCE block matrix structure
This commit includes a lot of cleanup but primarily is about the block
matrix structure necessary to solve the intrusive PCE system.  As such,
many of the meaningful changes in this commit are in the PCEBuilder
class.

As with many of the PCE objects, the PCE Builder class was originally a
copy of the embedded sampling (ES) builder, and I've now modified it to
support PCE.  In the ES builder, the block structure was a block
diagonal, and the blocks were independent of each other.  So, the system
was essentially a bunch of independent circuit-matrix-sized linear
systems.    Also, the number of blocks along the diagonal was equal to
the number of sample points (or the number of quadrature points, if
doing projection-based PCE).

For the PCE system, the number of blocks is not based on the number of
quad points.  For intrusive PCE, the solvers are solving for
coefficients of the PCE expansion, for every solution variable.  So, the
number of blocks is equal to the number of terms of the expansion.
Which may be less, or more than the number of quad points.  In some
cases the number of quad points and the number of expansion terms will
be the same, but not usually.

Quadrature is used by PCE to evaluate all the inner product integrals
numerically.  So, the capability to create block objects that have a
block size equal to the number of quadrature points is still necessary.
It just isn't used to create the linear system used by the nonlinear and
linear solvers.

So, one thing this commit does is to implement functions for both block
systems.  The default functions (which have names like "createMatrix"
and "createVector") create block objects based on the PCE expansion, as
these are needed by the NL solver.  The non-default functions (which
have names like "createQuadMatrix" and "createQuadVector") are also
implemented create block objects based on the number of quad points.
These "quad" functions will mainly be used by the PCELoader, and
probably nobody else.

This commit also fleshes out the function "generateGraphs" to set up the
block matrix structure for both the PCE and Quad matrices, as well as
all the various "generateMaps" functions.