Allow STEPNUM Column in Tabular Output Formats.
This commit adds support for .OPTIONS OUTPUT ADD_STEPNUM_COL,
which provides the capability to output a STEPNUM column, as the first
column, in FORMAT=STD, FORMAT=NOINDEX, FORMAT=GNUPLOT and FORMAT=SPLOT
output.  The STEPNUM column is not output for FORMAT=CSV, just as
the Index column is not in that case.

This commit also makes -remeasure work for TRAN and AC measures,
if the re-measured .prn file has a STEPNUM column.  The DC -remeasure
case still needs work.

SON Bug 1209
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1209