remove Xyce_FLEXIBLE_DAE_LOAD ifdef
The FLEXIBLE DAE LOAD ifdef is a holdover from MPDE, and was basically
an attempt to separate the residual and matrix loads into separate
function calls.  This is what is normally done when Xyce is run for DC
or transient, but it is not as easy to do when doing a block method,
such as MPDE or PCE.

This holdover continues to appear in the HB loader, and I then copied it
into the Embedded sampling loader, and then most recently into the PCE
loader.    I'm pretty sure it isn't used at all anymore, and maybe was
never really used.   But, this commit is only about PCE analysis, so I'm
not touching the ifdef anywhere else.

As PCE is new, and I'm still sorting thru code, I decided it would be
easiest if I de-cluttered this file.  So, I've removed the "flexible
DAE" code from it, as it was probably not ever going to be used.