Removes last vestiges of ground node in Jacobian graph/matrix
With the removal of the ground node from the Jacobian graph and matrix
there can be a clean up in the builder, matrix, and direct solvers.
The EpetraExt view transform that was used in the serial builds to view
the matrix without the ground node can be removed.  The generation of
the graph with the ground node included, can be removed and no longer
stored.  Finally, the matrix copy in the Amesos/Amesos2 interface can
be removed, as it is no longer necessary.