Add missing return to processParams
Clang is really screwy when a function is declared as returning a
value but doesn't actually return anything.  When the function exits
and hasn't returned a value, even if nobody that calls it USES the
return value, the resulting code does strange things --- this seems to
vary from Clang version to Clang version, and ranges from crashing
with an illegal instruction do doing what it did today, throwing
bizarre memory errors.

Instance::processParams in the new YLin device didn't return true, and
when I ran it on BSD with Clang, I got strange errors about the
multiplicity factor not being initialized before its use in a
conditional, and an out of memory crash --- despite it clearly having
been initialized, and despite that conditional being INSIDE the
processParams function and NOT obviously dependent on the return value
of the function.  But making the function return true fixes both the
valgrind errors *and* the weird out of memory crash, so I'm going
ahead and committing it.