Remove support for >20-year-old C standards
In working through the CMake refactor, we came across a bit of code in
the ksparse package. It was including "values.h", "limits.h" and/or
"float.h", depending on whether the build system found those files. The
"values.h" file was depricated prior to the C99 standard. Since I don't
see us trying to support standards that are over twenty years old, I
think it's safe to remove the ifdef's.