Make configure fail if required headers not found
In a recent commit, Jason removed conditionals around includes of
files that can reasonably be assumed to exist in a c99 compliant
compiler.  However, configure was still testing for them and setting
the macro that indicates their presence.

I could have removed the check altogether, but in lieu of that I am
instead making configure abort if it fails to find them, so that
someone trying to build Xyce as an experiment in retrocomputing will
get their disappointment at configure time instead of compile time.