Lots of work on the PCE loader
This commit does a lot of the work on the PCE loader.  I am committing
this as a checkpoint commit, as I am nearly done for the day.

The initial implementation of intrusive PCE is nearly done, I think.  The
loadDAEVectors function, which does 99% of the work in this class, now
has completely implemented all the vector stuff (converting to and from
the quadrature points and the expansion as needed).  The matrix stuff is
started, but not finished yet.

I had to add a new function to the data store class, called
"allocatePCEVectors", which forces the data store class to delete and
re-allocate the various vectors that need to have a block size equal to
the number of quad points.    The reason this was necessary is because
the solution vectors are kept at a block size of the size of the
expansion. So, unlike most other analysis types, some block vectors are
one block size and others are another.

This commit also includes a lot of cleanup, including to the embedded
sampling class.  There were some artifacts left over from HB that didn't
need to be there.