Fixes returnRawEntryPointer method to be more robust
Silly me, I was thinking that ground node LIDs would be -1 in the row
or column.  There are some special devices that come up with LIDs that
are not valid (in that they are ground) anymore.  There are some ints
and pointers that are not initialized in this method presuming that only
valid LIDs would enter that section of the code.  That is not true.
This may well be causing our issues with valgrind and bad device behavior.

I'm committing this change, because it is the right thing to do for a
method that should be erradicated.  Let's see what happens.