Add configure tests for boost headers
Since ATHENA's dependence on boost compiled libraries was removed,
Xyce's configure-time test for those libraries was removed.

Unfortunately, configure was never set up to test for the presence of
boost HEADERS that have always been required by ATHENA, and still
are.  That was OK when we were still checking for libraries, because
they always went together.  We have had several problems since then
with people at other labs trying to build XyceRad on systems that
didn't have boost, and finding out only at compile time that
something was wrong.

This commit adds configure tests for a few of the headers that ATHENA
needs, and if they are not found we just disable ATHENA and warn the
user with a message that explains how to fix it.

I have tested this by removing boost headers from my system and
confirming that the build is then set up to skip ATHENA just as if
"--disable-athena" were selected (but with a loud warning).  Reinstalling
the headers and rerunning configure gets ATHENA into my build again.

https://morannon.sandia.gov/bugzilla/show_bug.cgi?id=2125