Latest update on intrusive PCE
This commit includes the application of the triple prodct tensor to
creat the stochastic Jacobian.  This is work that is mostly in the PCE
loader.

It still has a bug in it (memory problem) but I need to save my work.
Hence, this commit.

Along the way, I did a few more things.  I chose to have the stochastic
block Jacobian be dense for now (possibly indefinitely).  I also started
to work on the PCEDirectSolver class.  I initially did this b/c I
thought maybe it would simply be easier to set up the some aspects of
the stocahstic system in that class.  But, about mid-way thru I realized
that what I had in mind wasn't going to work.  But most of my changes
to that class were still valid, so I'm am including them in this commit.