Modified getBlockAnalysisFlag to include PCE and ES
This is a very old function that I apparently wrote back in 2006.  It
was returning true only for HB and MPDE, but assuming that the issues
are similar, it should also be applied to the other 2 block analyses PCE
and EmbeddedSampling (ES).  I haven't rigorously tested this yet, but it
may be important for things like IC, if my memory serves.