Fixed mistake in applying triple product tensor
The tensor needed to be scaled by inverse basis norms, because when I
compute the projections used in the PCE residual, I am using that
scaling there.  So this needed to be include in the matrix setup for
consistency.

This commit also includes some very minor cleanup. It removes some
commented out code, and also comments out some debug print statements