Add .OPTIONS OUTPUT PHASE_OUTPUT_RADIANS
This commit adds .OPTIONS OUTPUT PHASE_OUTPUT_RADIANS=<boolean>.
This new option allows the Xyce phase output, for the VP(), IP(),
SP(), XP() and ZP() operators and for .PRINT SENS for a .AC
analysis to be toggled between being output in radians vs. degrees.
The default is still for phase output in radians.  The phase angles
in Touchstone files are still only output in degrees, per that
standard.

For the operators, the technical approach was to make separate
operator classes, where the eval() function for each class now
returns radians or degrees.  The choice of which operator gets made
is determined by the PHASE_OUTPUT_RADIANS setting.

There are no operators currently defined for Sensitivity output.
So, the conversion between radians and degrees is done within the
AC object, and its associated free functions, rather than in
the outputter classes like SensitivityACPrn.

SON Bug 1208
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1208