Improve Error Handling for .PCE Analyses
This commit adds a test for the error message, for a stokhos
build, when the ORDER parameter on the .OPTIONS PCE line has an
invalid value.  For all builds, it tests for illegal values for
ALPHA, BETA and STD_DEVIATIONS. For the gamma distribution, alpha
and beta must be strictly positive.  A standard deviation must be
non-negative for any distribution.

SON Bug 1224
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1224