Initial Capability for PCE Outputters
This is the initial capability for the outputters for Polynomial Chaos
Expansions (.PCE).  Features of the new .PRINT PCE line, which is only
compatible with TRAN and DC analyses are as follows:

  1) FORMAT=<STD|NOINDEX|GNUPLOT|SPLOT> will make output files of the
form <netlistName>.PCE.prn.

  2) FORMAT=CSV will make output files of the form <netlistName>.PCE.csv.

  3) FORMAT=TECPLOT will make output files of the form <netlistName>.PCE.dat.

  4) FORMAT=<RAW|PROBE|TOUCHSTONE|TOUCHSTONE2> will default to FORMAT=STD.

  5) The column list in each output file includes those that are generated
by the function PCE::hackPCEOutput().  The columns can also include data
from all of the Quadrature Points if outputAllSamples=true is included
on the .PRINT PCE line.

  6) A request for -r or -a, when .PCE is used in the netlist, is
currently a parsing error.

  7) For a -o <fileName> command line option, the output from the .PRINT PCE
line will be sent to <fileName> in FORMAT=STD.

  8) The output from the .PRINT TRAN or .PRINT DC lines in the netlist
will occur, if -o is not specified on the command line.  (Note: this may
be changed in the future.)

  9) No output will occur from the .PRINT PCE line if the .OPTIONS PCES
line does not have an OUTPUTS= argument.

SON Bugs 1229 and 1231
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1229
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1231