Implement voltage limiting for intrusive PCE
This is in support of SON bug 1229, "Implement the fully intrusive
form of polynomial chaos expansions (PCE) in Xyce"

http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1229

Prior to today I was not very successful in getting nonlinear problems
to run properly with intrusive PCE.  Part of the problem was that I was
scaling the matrix incorrectly, which I fixed on a previous commit.  The
other part of the problem was that I had not implemented voltage
limiting in a way consistent with PCE.

I have done that now.   The method requires that additional
circuit-sized linear systems be solved for each quadrature
point, whenenver the voltage limiter vectors have a nonzero max norm.
The linear solve is of this system:

  dFdx * dx_vl = dFdxdVp

Once dx_vl is obtained for each quadrature point, then the pce expansion
for it can be computed.  Then, after we also have the stochastic
jacobians for F and Q, we can compute the new voltage limiter vectors
using matvecs.   These limiter vectors are then added to the DAE system
to be solved.