Outputter Changes for .SAMPLING, .PCE and .EMBEDDEDSAMPLING
This commit changes the output format for .SAMPLING, .PCE and
.EMBEDDEDSAMPLING so that the variables in the OUTPUTS= and
MEASURES= arguments, from their associated .OPTIONS lines, are
upper-cased. So, for .PRINT ES, the argument outputs={v(1)},{V(2)}
would produce columns with the strings {V(1)}_mean and {V(2)}_mean.
The OUTPUTS variables are upper-cased, while the suffix-text
associated with the ES analysis is still lower-cased.  The format
change for MEASURES= only applies to .SAMPLING.

This commit also changes the format for OUTPUT_PCE_COEFFS=true for
.EMBEDDEDSAMPLING. An underscore is now used as the "separator"
since commas interfered with reading FORMAT=CSV files into Excel.
An example of the new format is {V(1)}_coef(0_0).

SON Bugs 1201 and 1231
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1201
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1231