Fixed a memory error in intrusive PCE
Intrusive PCE requires that some block vectors have the number of blocks equal to
the number of quadrature points, and other block vectors (particularly the linear
system vectors/matrices) have the number of blocks equal to the number of PCE
coefficients.

This is a different than how most other block analyses work.  Most others expect
that all the block objects have the same number.

I initially got around this by having the PCEBuilder be able to create block
objects of both sizes, and adding the name "Quad" to functions the produced
Quad-sized block objects, and leaving the default name for functions producing
coef-sized objects.  Doing it this way meant that the DataStore class then
allocated everything to be coefficient sized, and this had to be corrected
later, with certain objects being deleted and reallocated to be quad sized.

All of this got to be a bit too clever and caused at least one memory mistake
that was caught by valgrind.

I've now refactored the PCEBuilder and its use in the data store object to be
simpler and cleaner, and the memory error is gone.