Fix 2 mistakes in PCE voltage limiting
This commit is in support of SON bug 1229, "Implement the fully intrusive
form of polynomial chaos expansions (PCE) in Xyce":
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1229

As the summary says, this commit fixes two mistakes.

Mistake #1:
When I implemented PCE voltage limiting the other day, I got it working
first, and then attempted to make it more efficient.  The final commit
included both the "got it working" code as well as the "make it more
efficient" code.  It should have only included the "got it working"
code, I now realize.

The "make it more efficient" idea was to look at the magnitude of the
limiter vectors that were returned from loadDAEVectors at each
quadrature point, and for the ones that were zero, then don't bother doing
anything.  Simple idea, but I made a mistake in it.  So, commenting that
out for now.  Now, the volt lim code is applied at every step, as long
as voltage limiting is enabled in the device pacakge.  I have a diode
test circuit that runs great as long as the "make it more efficient"
code is commented out.

Mistake #2:
When I solve the voltlim linear system at each
quadrature point, I was making the mistake of setting the "reuse LU
factors" boolean to "true".  This should only be true if you are
re-using the same matrix and that is not generally the case here.  So it
was a mistake and is now fixed.

With this fix, the solver is working even better.  I have a CMOS inverter
circuit that, before this fix, ran fine if I only had one uncertain
parameter, but failed to converge with two uncertain parameters.  Now it runs
fine even with two.  I had initially thought that maybe I was screwing
up the multivariate polynomial somehow, but that turned out not to be
the problem, fortunately.

For both of the above circuits (the simple diode and the CMOS inverter),
if ".options device voltlim=0" (which means that the device package
will not generate voltage limiter terms at all), they fail
to converge.  If I set "voltlim=1", but turn off the PCE
voltlim algorithm by setting ".options PCES voltlim_alg=0", then the diode
circuit converges but takes 30 or so iterations, and the CMOS inverter
fails.  If I have both ".options device voltlim=1" and ".options PCES
voltlim_alg=1", then both circuits converge.  The diode circuit requires
3 iterations.  The CMOS circuit takes around 8 iterations for the DCOP
and then successfully simulates all the way thru the transient.

So, it really matters for nonlinear circuits to have voltage limiting on
(we knew this already), and it also really matters to have the voltlim
terms be mathematically consistent with PCE.