Code Clean-up for .PRINT ES and .PRINT PCE
This commit refactors the outputters for .PRINT PCE and .PRINT ES
so that common functions, used by the PCEPrn, PCECSV and PCETecplot
classes to output header info and data rows, have been moved to a
common set of functions that should be easier to maintain.  A similar
refactor was done for the EmbeddedSamplingPrn, EmbeddedSamplingCSV and
EmbeddedSamplingTecplot classes.

SON Bugs 1201 and 1231
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1201
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1231