Corrected the addDevicesConverged for PCE and ES
The nonlinear solver has, as part of its convergence tests, a function
call that ultimately goes to the device package asking if all the
devices in the circuit are considered "converged".  Mainly this is to
ensure that the nonlinear solver doesn't stop iterating when the devices
are still applying voltage limiting.  This can occasionally happen
otherwise, as the norms may all be small but the devices aren't quite
there yet.

This call has always, regardless of analysis type, gone straight thru
the NonlinearEquation loader to the device package, without any
analysis-specific loaders in between.  I think this was technically
incorrect for block-based nonlinear solves.  If doing a block-based
solve, such as ES, the value returned would only represent the most
recent block (or sample point) that was evaluated.  This is always
correct for linear problems (as the function always returns true in that
case anyway), and often would be pretty close to correct for a lot of
nonlinear problems.  But not 100%.  For cases in which some blocks have
all devices converged and some blocks do not, this was wrong.

I've modified the loaders so that now "allDevicesConverged" is a purely
virtual function and must be implemented.  It had already been
implemented in the CktLoader and most analyses (including ES, PCE and
HB) derive from it.  It had not been implemented in any of the MPDE
loaders before and also had not been implemented for ES or PCE.
I have now implemented it for ES,PCE and MPDE.

I did not touch HB, as I ran out of time and brain cycles today,
but this commit should preserve its previous behavior.

For the 3 modified analyses types (ES,PCE and MPDE) this will make the
nonlinear solver slightly more conservative, and also slightly more
robust.  It probably won't make a huge change in solver behavior.