Removes unnecessary generation of solution coloring maps
Epetra_MapColoring object is being created in the N_LAS_System object
even thought it is never used by anything else in Xyce.  The nonlinear
solver requests that the Epetra_MapColoring object be created if gmin
stepping or initial conditions are being set.  So, this additional map
can be removed from the N_LAS_System object.  Furthermore, the size of
the linear system is being obtained by the N_LAS_QueryUtil object when
it could be obtained from the N_LAS_Builder object.  This has been
changed also so that the requirement for the N_LAS_QueryUtil object
can be removed from the N_LAS_System.