Moves access of vnodeGIDVec to builder from system
The linear system is agnostic of the dimension of the problem being
solved.  This is information that is in the builder, that the linear
system uses to generate the necessary matrices and vectors.  Furthermore,
the N_LAS_QueryUtil object is only valid for transient analysis types,
some of the information obtained from it are not correct for block
analysis types.  It's not necessary for the N_LAS_QueryUtil object to
be modified for block analysis if those pieces of information are being
obtailed through a builder.

So, this commit removes the N_LAS_QueryUtil object from the system,
as well as the accessor provided by the system.  The builder is assuming
some of the duties required for .IC and gmin stepping so it will be
correct for block analysis methods like HB and embedded sampling.