Small fixes and revisions to Sampling, ES and PCE
I was working today on examples for my final SAND report, and noticed a
few small things that needed fixing.  They are all included in this
commit.

(1) the OUTPUTSAMPLESTATS parameter, for sampling, ES and PCE was
causing conflicts with the OUTPUTS parameter, as the first 7 characters
were the same.  I've changed it to OUTPUT_SAMPLE_STATS so that conflict
is gone.

(2) Sampling analysis needed a "Starting sampling simulation" banner, as
well as output explaining how many quadrature points were being used in
the case of PCE.

(3) For the example I was working on, I wanted to output mean+/-2*sigma.
Up until now I could only output mean+/-1*sigma.  I've added it, but
only to the "Hack" output functions.