Fix logic for user warning for using measure w/PCE
The PCE methods allow you to use .measures as the output function of
interest.  In all 3 of these methods, I added a warning that would tell
the user when a specified measure was "not found" by the measure
manager.  However, I didn't get the logic right; it was emitting the
warning unconditionally.  And, as I copied this code from sampling to
embededsampling and then to pce, I had to fix it in 3 places.

I didn't notice it before now b/c I haven't used .measure with pce all
that much yet, and I'm usually running a verbose build so this warning
was getting lost in the noise.