Removes Epetra_MapColoring usage from builder and homotopy/.IC/.NODESET
The Epetra_MapColoring class is too heavy weight for this simple
coloring of the solution variables.  The std::vector<int> object
created by the builder classes is now being used to implement the
gmin stepping, .IC, and .NODESET approaches.  This was, in effect,
how the Epetra_MapColoring class was being used, as a std::vector<int>.

This also reduces the dependencies on Epetra-based objects, which
is a benefit in the long run.