Undo "optimization" applied for large tables
In commit c9a67d88 I made the expression package skip doing common
subexpression elimination (CSE) when the current subexpression under
consideration is just a constant.  Doing so led to a significant
speed-up in the parsing of large tables created using the "TABLEFILE"
feature.

It was, unfortunately, the wrong thing to do.  Yes, it sped up that
use case, but it also broke a lot of potential CSE cases (resulting in
inefficiencies in expression evaluation in some cases).

The correct fix for the tablefile bottleneck lies elsewhere, as
documented in bug 1176, bug 1222, and bug 1241.

I am therefore undoing the change.  While I was at it, I cleaned up
some blank lines and comments.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1241
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1222
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1176