Add functions to convert between S and Y parameters
The S and Y parameter conversions are required in both S-parameter
analysis and S-parameter models. The Y to S conversion is needed
for SP analysis. The S to Y is needed in order to read in S parameters
as a device. The Y to S conversion is rewritten with a differnt
formula which may be slightly more efficient. The S to Y uses
a matrix format conversion. It can be used in a general frequency
domain model, such as S-parameter device